Arm Programming
A lot of arm movement was done, we still need to add encoder movement for the turntable, but for now this is good. The two servos on the arm move to position well, however servo2 probably needs to be updated because it's range of movement is very off and weird. Other than that we are in very good shape.